Tunables refactoring: Config-related code to db/config.c
